refactor(guest-flow): extract confirmation logic into custom hook
Moved guest confirmation flow logic from ConfirmChildrenFlowPage into a dedicated hook (`useGuestConfirmationFlow`) to improve separation of concerns and code reusability.

Also:
- Updated ConfirmChildrenFlowPage to use the new hook
- Minor formatting and cleanup in related components
